+++ project/api/tests/__init__.py
@@ -1,3 +1,12 @@
 """Module that contains tests"""
 from .model_tests import *
-from .api_tests import *
+from .cd_tests import *
+from .city_tests import *
+from .ms_tests import *
+from .narration_tests import *
+from .narrative_tests import *
+from .narrative_vote_tests import *
+from .pr_tests import *
+from .profile_tests import *
+from .stv_tests import *
+from .te_tests import *
