+++ project/api/tests/cd_tests.py
@@ -1,0 +1,79 @@
+from rest_framework import status
+from django.urls import reverse
+from api.models import CachedData
+from .api_tests import APITest
+
+class CDTests(APITest):
+    """
+    Cached Data test suite
+    """
+
+    def test_api_can_create_cd(self):
+        """
+        Ensure we can create CachedData
+        """
+
+        url = reverse("cacheddata-list")
+        data = {
+            "wikidata_id": 2,
+            "location": "Point(0 1)",
+            "date": self.JD_0001,
+            "event_type": CachedData.DOCUMENT,
+        }
+        response = self.client.post(url, data, format="json")
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+        self.assertEqual(CachedData.objects.count(), 2)
+        self.assertEqual(CachedData.objects.last().event_type, CachedData.DOCUMENT)
+
+    def test_api_can_create_cd_othertype(self):
+        """
+        Ensure we can create CachedData with an event_type not in the choices
+        """
+
+        url = reverse("cacheddata-list")
+        data = {
+            "wikidata_id": 2,
+            "location": "Point(0 1)",
+            "date": self.JD_0001,
+            "event_type": 555,
+        }
+        response = self.client.post(url, data, format="json")
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+        self.assertEqual(CachedData.objects.count(), 2)
+        self.assertEqual(CachedData.objects.last().event_type, 555)
+
+    def test_api_can_update_cd(self):
+        """
+        Ensure we can update CachedData
+        """
+
+        url = reverse("cacheddata-detail", args=[self.hastings.pk])
+        data = {
+            "wikidata_id": 1,
+            "location": "Point(0 0)",
+            "date": self.JD_0001,
+            "event_type": CachedData.DOCUMENT,
+        }
+        response = self.client.put(url, data, format="json")
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        self.assertEqual(response.data["event_type"], CachedData.DOCUMENT)
+
+    def test_api_can_query_cds(self):
+        """
+        Ensure we can query for all CachedDatas
+        """
+
+        url = reverse("cacheddata-list")
+        response = self.client.get(url, format="json")
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        self.assertEqual(response.data[0]["event_type"], CachedData.BATTLE)
+
+    def test_api_can_query_cd(self):
+        """
+        Ensure we can query for individual CachedDatas
+        """
+
+        url = reverse("cacheddata-detail", args=[self.hastings.pk])
+        response = self.client.get(url, format="json")
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        self.assertEqual(response.data["event_type"], CachedData.BATTLE)
