+++ project/api/tests/narration_tests.py
@@ -1,0 +1,70 @@
+from rest_framework import status
+from django.urls import reverse
+from api.models import Narration
+from .api_tests import APITest
+
+class NarrationTests(APITest):
+    """
+    Narration test suite
+    """
+
+    def test_api_can_create_narration(self):
+        """
+        Ensure we can create Narrations
+        """
+
+        url = reverse("narration-list")
+        data = {
+            "narrative": self.norman_conquest.pk,
+            "title": "Test Narration",
+            "description": "This is a narration point",
+            "date_label": "test",
+            "map_datetime": self.JD_0002,
+            "settings": self.norman_conquest_settings.pk,
+            "attached_events_ids": [self.hastings.pk],
+            "location": "POINT (0 0)",
+        }
+        response = self.client.post(url, data, format="json")
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+        self.assertEqual(Narration.objects.count(), 2)
+        self.assertEqual(Narration.objects.last().title, "Test Narration")
+
+    def test_api_can_update_narration(self):
+        """
+        Ensure we can update Narrations
+        """
+
+        url = reverse("narration-detail", args=[self.hastings_narration.pk])
+        data = {
+            "narrative": self.norman_conquest.pk,
+            "title": "Test Narration 2",
+            "description": "This is a narration point",
+            "date_label": "test",
+            "map_datetime": self.JD_0002,
+            "settings": self.norman_conquest_settings.pk,
+            "attached_events_ids": [self.hastings.pk],
+            "location": "POINT (0 0)",
+        }
+        response = self.client.put(url, data, format="json")
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        self.assertEqual(response.data["title"], "Test Narration 2")
+
+    def test_api_can_query_narrations(self):
+        """
+        Ensure we can query for all Narrations
+        """
+
+        url = reverse("narration-list")
+        response = self.client.get(url, format="json")
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        self.assertEqual(response.data[0]["title"], "Test Narration")
+
+    def test_api_can_query_narration(self):
+        """
+        Ensure we can query for individual Narrations
+        """
+
+        url = reverse("narration-detail", args=[self.hastings_narration.pk])
+        response = self.client.get(url, format="json")
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        self.assertEqual(response.data["title"], "Test Narration")
