+++ project/api/tests/narrative_tests.py
@@ -1,0 +1,76 @@
+from rest_framework import status
+from django.urls import reverse
+from django.test import tag
+from api.models import Narrative
+from .api_tests import APITest
+
+class NarrativeTests(APITest):
+    """
+    Narrative test suite
+    """
+
+    def test_api_can_create_narrative(self):
+        """
+        Ensure we can create Narratives
+        """
+
+        url = reverse("narrative-list")
+        data = {
+            "author": "Test Author 2",
+            "title": "Test Narrative",
+            "url": "test2",
+            "description": "This is a test narrative for automated testing.",
+            "tags": ["test", "tags"],
+        }
+        response = self.client.post(url, data, format="json")
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+        self.assertEqual(Narrative.objects.count(), 2)
+        self.assertEqual(Narrative.objects.last().author, "Test Author 2")
+
+    def test_api_can_update_narrative(self):
+        """
+        Ensure we can update Narratives
+        """
+
+        url = reverse("narrative-detail", args=[self.norman_conquest.pk])
+        data = {
+            "author": "Other Test Author",
+            "title": "Test Narrative",
+            "url": "test2",
+            "description": "This is a test narrative for automated testing.",
+            "tags": ["test", "tags"],
+        }
+        response = self.client.put(url, data, format="json")
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        self.assertEqual(response.data["author"], "Other Test Author")
+
+    def test_api_can_query_narratives(self):
+        """
+        Ensure we can query for all Narratives
+        """
+
+        url = reverse("narrative-list")
+        response = self.client.get(url, format="json")
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        self.assertEqual(response.data[0]["author"], "Test Author")
+
+    def test_api_can_query_narrative(self):
+        """
+        Ensure we can query for individual Narratives
+        """
+
+        url = reverse("narrative-detail", args=[self.norman_conquest.pk])
+        response = self.client.get(url, format="json")
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        self.assertEqual(response.data["author"], "Test Author")
+
+    @tag("new")
+    def test_api_can_query_narrative_votes(self):
+        """
+        Ensure upvotes and downvotes are being returned correctly
+        """
+
+        url = reverse("narrative-detail", args=[self.norman_conquest.pk])
+        response = self.client.get(url, format="json")
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        self.assertEqual(response.data["votes"], {"upvotes": 1, "downvotes": 0})
